Resolve unnecessary mathesar_db container startup in dev-service
Description:
This pull request addresses the issue #2973 reported by @pavish. The problem is related to the unnecessary startup of the mathesar_db container when running the dev-service.

To resolve this issue, the following changes have been made to the Docker Compose configuration files:

In the docker-compose.yml file:
Renamed the mathesar_db service to dev-db.
Removed the unnecessary mathesar_db container references from the test-service and dev-service sections.
These changes ensure that only the necessary containers (mathesar_dev_db and mathesar_service_dev) are started when running the dev-service, eliminating the unused mathesar_db container.

Please review and test this pull request to confirm that the issue is resolved as expected.

Note: If there are any additional considerations or requirements for the mathesar_db container, please provide the necessary details so that the configuration can be adjusted accordingly.

Related Issue: #2973